--- a/contributors.yml
+++ b/contributors.yml
@@ -102,6 +102,7 @@
 - dhargitai
 - dhmacs
 - dima-takoy
+- dmarkow
 - DNLHC
 - dogukanakkaya
 - dokeet