Closes: #4480. `defineConventionalRoutes` was repeatedly running `.filter` and `.find` against the list of routeIds to find parents, which was causing an exponential slowdown in larger projects. This update first generates a `parentRouteIds` object and then filters against that instead.

- [ ] Docs
- [x] Tests

Testing Strategy: There weren't any direct tests for `defineConventionalRoutes` as far as I could tell, so I added one to make sure the necessary number of routes (including parent route ids) are generated. I used an existing fixture for the routes folders (`remix-dev/__tests__/fixtures/replace-remix-imports`) but I can add one specific to the `routesConventions` tests if you prefer this test remain uncoupled from the replace imports test.

I ran this against my production app with nearly 700 routes and the RouteManifest object was was identical before and after the patch. Before this update, my production app was taking around 10-15 seconds before "Watching Remix app in development mode..." showed up, now it's less than a second.

I also ran the [example test app from my original issue](https://github.com/dmarkow/remix-route-performance-example) against this new code and the speed up is drastic on large projects (almost 27 seconds before for 1,111 routes, now it's 0.104 seconds).

```
$ node run.js 1 | head
Creating sample routes...
defineConventionalRoutes: 0.008s for 112 routes (0.07ms per file)
<Routes>
  <Route file="root.tsx">
    <Route path="0/10/0/0" index file="routes/__a/__b/__c/0/10/0/0/index.js" />
    <Route path="0/10/0/1" index file="routes/__a/__b/__c/0/10/0/1/index.js" />
    <Route path="0/10/1/0" index file="routes/__a/__b/__c/0/10/1/0/index.js" />
    <Route path="0/10/1/1" index file="routes/__a/__b/__c/0/10/1/1/index.js" />
    <Route path="0/10/2/0" index file="routes/__a/__b/__c/0/10/2/0/index.js" />
    <Route path="0/10/2/1" index file="routes/__a/__b/__c/0/10/2/1/index.js" />

$ node run.js 5 | head
Creating sample routes...
defineConventionalRoutes: 0.044s for 556 routes (0.08ms per file)
<Routes>
  <Route file="root.tsx">
    <Route path="0/10/0/0" index file="routes/__a/__b/__c/0/10/0/0/index.js" />
    <Route path="0/10/0/1" index file="routes/__a/__b/__c/0/10/0/1/index.js" />
    <Route path="0/10/1/0" index file="routes/__a/__b/__c/0/10/1/0/index.js" />
    <Route path="0/10/1/1" index file="routes/__a/__b/__c/0/10/1/1/index.js" />
    <Route path="0/10/2/0" index file="routes/__a/__b/__c/0/10/2/0/index.js" />
    <Route path="0/10/2/1" index file="routes/__a/__b/__c/0/10/2/1/index.js" />

$ node run.js 10 | head
Creating sample routes...
defineConventionalRoutes: 0.104s for 1111 routes (0.09ms per file)
<Routes>
  <Route file="root.tsx">
    <Route path="0/10/0/0" index file="routes/__a/__b/__c/0/10/0/0/index.js" />
    <Route path="0/10/0/1" index file="routes/__a/__b/__c/0/10/0/1/index.js" />
    <Route path="0/10/1/0" index file="routes/__a/__b/__c/0/10/1/0/index.js" />
    <Route path="0/10/1/1" index file="routes/__a/__b/__c/0/10/1/1/index.js" />
    <Route path="0/10/2/0" index file="routes/__a/__b/__c/0/10/2/0/index.js" />
    <Route path="0/10/2/1" index file="routes/__a/__b/__c/0/10/2/1/index.js" />
```