@dmarkow 2022-11-07
@machour My git rebasing skills are apparently trash 😅 Rebasing my local branch against dev is still keeping all of main's recent commits as well, and changing the base branch of this PR causes 11 different commits from main to pull over.

Can you point me in the right direction on this, or should I just start over and push a new PR originating from the dev branch?